choose int to short

+++ gibumfile/gibumfile.cpp
@@ -5,7 +5,7 @@ using namespace std;
 namespace fs=std::filesystem;
 int main(){
     while(1){
-    int choose;
+   short choose;
     cout<<"다음중 선택세요(1.파일명 일괄 변경/2.특정 문자 지우기):";
     cin>>choose;
     if( choose==1){
